Export: use the Compositor for faster image composition
The unhappy path in ExportImage previously did manipulation on
the image.pixels field to create an appropriate image out of the
channels of other images.

This changes it to create a temp scene and setup the Compositor in
that scene to create the necessary image, then render. Despite
being pretty hacky, this is about 4 times faster (ie. it does 4
times more pixels per second).

This also requires knowing what colorspace the target image should
be in.